Refactor: Remove env and have consistent naming
