--- .env
@@ -1,1 +1,0 @@
-REACT_APP_JWT_SECRET=reactapp
