+++ public/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <title>VelStream</title>
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <link rel="stylesheet" href="https://velocity-practice.netlify.app/css/components.css">
+    <link rel="stylesheet" href="https://velocity-ui-dev.netlify.app/css/components.css">
     <script src="https://kit.fontawesome.com/3582c9ec8a.js" crossorigin="anonymous"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
