+++ src/reducers/index.js
@@ -2,4 +2,4 @@ export { authReducer, initialAuthState } from "./authReducer";
 export { likeReducer, initialLikeState } from "./likeReducer";
 export { playlistReducer, initialPlaylistState } from "./playlistReducer";
 export { watchLaterReducer, initialWatchLaterState } from "./watchLaterReducer";
-export { historyReducer, initialHistoryState } from "./history-reducer";
+export { historyReducer, initialHistoryState } from "./historyReducer";
